% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_schedule.R
\name{cross_scheduler}
\alias{cross_scheduler}
\title{schedule your SGA crosses}
\usage{
cross_scheduler(duration, crosses, write = T)
}
\arguments{
\item{duration}{path to a csv file specifying, on each row: each step of the crossing procedure and the number of days between one given step and the first liquid culture launching step.}

\item{crosses}{path to a txt file (tab delimited) specifying the id (has to be unique) of the cross, the queries in the cross (separated by commas), and the date (YYYY-MM-DD) at which liquid cultures were launched before making lawns.}
}
\description{
schedule your SGA crosses
}

